对照open ai 官网添加注释
