+++ src/Thor.Abstractions/Chats/Consts/ThorToolTypeConst.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Thor.Abstractions.Chats.Consts
+{
+    /// <summary>
+    /// 工具类型定义
+    /// </summary>
+    public class ThorToolTypeConst
+    {
+        /// <summary>
+        /// 函数
+        /// </summary>
+        public static string Function => "function";
+    }
+}
